Introduce Repository (#5, #24)
- Introduce Repository class which handles remote package repositories,
  responsible for (yes, there's some SRP violation smell):
  - Fetching metadata from the net.
  - Parsing and storing it (in a more compact format).
  - Returning relevant package information by name, origin or
    name+version.
- Introduce RepositoryManager which dispatches to a right repository
  based on jail release and arch.
- It's now possible to request a binary package anywhere in the build
  and it will fetched and returned as an object hondling metadata and
  absolute path.
- Use it to bootstrap pkg in the jail. It eliminates network access
  for each build.

In the future, we can use the Repository for dependency walking and
for installing binary packages, eliminating a need for a lot of `pkg`
calls and a long `pkg update` for each build.

For now, all package metadata is stored in the memory, and it takes
around a dozen megabytes per repository. This memory can be conserved
by using local sqlite at some point, as it'll become more relevant
when reprise daemon is implemented, which would have to hold information
for all available jails.