Implement ccache support (fixes #2)
We rely on ports for handling ccache dependency and setting the
right CC/CXX. All we basically need is to enable WITH_CCACHE_BUILD
set CCACHE_DIR, and provide a persistent storage/mountpoint for it.

Notably, we use different ccache dirs for root and nobody builds to
avoid ownership problems.