Switch dependency walking to Repository (fixes #23)
As we now have all the required repository metadata, it's no longer
needed to call pkg to retrieve the same thing from its own database.
It is the first step to dropping the requirement for pkg database
completely. This significantly speeds up dependency grap calculation,
but introduces desynchrnization of two package databases which should
be gone in one of the following commits when we drop pkg database
completely.