Switch package installation to Repository (fixes #24)
We no longer rely on `pkg` for remote package installation,
so we may drop `pkg update` call completely, and switch PackageTask
to our Repository.

This is getting a bit ugly as the new architecture does not lay
well on Tasks, but this is expected and will be refactored later
when it's possible to reuse packages built by us. The target
architecture will not have Tasks, but will have a list of packages
to install (some of which may need to be built in different jobs),
which can also be done in one `pkg` call, and after packages
installation port build/install/test phases will be run explicitly.

Also fixes #4 and #5